pyproject.toml
[tool.poetry]
name = "citc"
version = "0.3.0"
description = ""
authors = ["Matt Williams <matt@milliams.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6"
boto3 = "^1.11"
PyYaml = "^5.3"
oci = "^2.9.0"
tabulate = "^0.8.6"
google-api-python-client = "^1.7.11"
google-auth-oauthlib = "^0.4.1"

[tool.poetry.dev-dependencies]
pytest = "^5.3"
pytest-mock = "^1.13.0"
black = {version = "*", allow-prereleases = true}
mypy = "^0.760"
flake8 = "^3.7.9"
moto = "^1.3.14"
coverage = "^5.0.3"

[tool.poetry.scripts]
watchdog = 'citc.watchdog:main'
list_nodes = 'citc.list_nodes:main'

[build-system]
requires = ["poetry>=1.0.0"]
build-backend = "poetry.masonry.api"
